splitnet/requirements.txt
scipy==1.7.1
matplotlib==3.4.3
torch==1.8.1
numpy==1.20.2
hydra-core
omegaconf==2.1.1
progressbar33==2.4
pytorch_lightning==1.4.9
rich==10.12.0
scikit_learn==1.0
seaborn==0.11.2
torchmetrics==0.5.1
dataclasses
